Menú creado, sin que se guarden los datos en ningún sitio
